SWARM-1932: MP fault tolerance: circuit breaker doesn't open if there was at least one success
Motivation
----------
The code in `SynchronousCircuitBreaker` used for figuring out
if the failure threshold of the circuit breaker has been reached
used to divide two `int`s and expect it will be a proper `double`.
That's obviously wrong in Java.

Modifications
-------------
Cast one of the `int`s to `double` so that we get proper division.

Result
------
The `SynchronousCircuitBreaker` is now better able to figure out
if the failure threshold has been reached.

+++ fractions/microprofile/microprofile-fault-tolerance/src/main/java/org/wildfly/swarm/microprofile/faulttolerance/deployment/SynchronousCircuitBreaker.java
@@ -164,7 +164,7 @@ class SynchronousCircuitBreaker implements HystrixCircuitBreaker {
         if (requestCount < requestVolumeThreshold) {
             return false;
         }
-        double failureCheck = failureCount.get() / requestCount;
+        double failureCheck = failureCount.get() / (double) requestCount;
         double failureRatio = config.get(CircuitBreakerConfig.FAILURE_RATIO);
         return (failureCheck >= failureRatio) || (failureRatio <= 0 && failureCheck == 1);
     }
